% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{as_wk_tibble}
\alias{as_wk_tibble}
\title{Convert raw fetch result to tibble with wk geometry}
\usage{
as_wk_tibble(raw, geom_col)
}
\description{
Convert raw fetch result to tibble with wk geometry
}
\keyword{internal}
