% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{build_wfs_dsn}
\alias{build_wfs_dsn}
\title{Build a WFS connection string for capabilities / layer listing}
\usage{
build_wfs_dsn(base_url, version = NULL, srs = NULL)
}
\description{
Build a WFS connection string for capabilities / layer listing
}
\keyword{internal}
