% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{build_wfs_read_dsn}
\alias{build_wfs_read_dsn}
\title{Build a WFS GetFeature DSN with bbox/count in the URL}
\usage{
build_wfs_read_dsn(
  base_url,
  layer = NULL,
  bbox = NULL,
  max_features = NULL,
  version = NULL,
  srs = NULL
)
}
\description{
Build a WFS GetFeature DSN with bbox/count in the URL
}
\keyword{internal}
