% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{detect_driver}
\alias{detect_driver}
\title{Detect the likely OGR driver from a URL}
\usage{
detect_driver(url)
}
\arguments{
\item{url}{Character. Service URL.}
}
\value{
Character. One of \code{"WFS"}, \code{"OAPIF"}, \code{"ESRIJSON"}.
}
\description{
Detect the likely OGR driver from a URL
}
\keyword{internal}
