% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{wfs_fields}
\alias{wfs_fields}
\title{Get field (column) information for a layer}
\usage{
wfs_fields(base_url, layer, driver = "auto", version = NULL, srs = NULL)
}
\arguments{
\item{base_url}{Character. The service endpoint URL.}

\item{layer}{Character. Layer name.}

\item{driver, version, srs}{Passed to \code{\link[=wfs_connection]{wfs_connection()}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns: \code{name}, \code{type}, \code{width},
\code{precision}, \code{is_nullable}.
}
\description{
Returns a tibble describing each non-geometry field in a layer:
name, type, width, and whether it's nullable.
}
\examples{
\dontrun{
url <- wfs_example_url("list_tasmania")
wfs_fields(url, "Public_OpenDataWFS:LIST_Cadastral_Parcels",
           version = "2.0.0", srs = "EPSG:28355")
}
}
